feat: Complete 01-LifeCycle
